% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuvpn.R
\name{wu_vpn_connect}
\alias{wu_vpn_connect}
\title{WU VPN Connect}
\usage{
wu_vpn_connect(
  user = Sys.getenv("WUSTL_KEY_USER"),
  pass = Sys.getenv("WUSTL_KEY_PASS"),
  vpn_cli = Sys.getenv("WU_VPN_CLI_PATH"),
  vpn_endpoint = Sys.getenv("WU_VPN_ENDPOINT")
)
}
\arguments{
\item{user}{wustl key username}

\item{pass}{wustl key password}

\item{vpn_cli}{path to Cisco AnyConnect VPNCLI.exe}

\item{vpn_endpoint}{vpn endpoint}
}
\description{
Connect to the WUSTL VPN using the Cisco AnyConnect VPN CLI client.
}
