% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuvpn.R
\name{wu_vpn_disconnect}
\alias{wu_vpn_disconnect}
\title{WU VPN Disconnect}
\usage{
wu_vpn_disconnect(vpn_cli = Sys.getenv("WU_VPN_CLI_PATH"))
}
\arguments{
\item{vpn_cli}{path to Cisco AnyConnect VPNCLI.exe}
}
\description{
Disconnect from the WUSTL VPN using the Cisco AnyConnect VPN CLI client.
}
