% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuvpn.R
\name{wu_vpn_status}
\alias{wu_vpn_status}
\title{WU VPN Status}
\usage{
wu_vpn_status(vpn_cli = Sys.getenv("WU_VPN_CLI_PATH"))
}
\arguments{
\item{vpn_cli}{path to Cisco AnyConnect VPNCLI.exe}
}
\description{
Check the status of the WUSTL VPN connection using the Cisco AnyConnect VPN
CLI client.
}
